% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_format_standard.R
\name{dt_format_standard}
\alias{dt_format_standard}
\title{Get a localized date-time in a standardized form.}
\usage{
dt_format_standard(input, locale = "en_US", width = "medium")
}
\arguments{
\item{input}{the input date-time. The appropriate representation should use
the following construction outlined in the ISO 8601:2004 standard:
\code{YYYY-MM-DDThh:mm:ss.sTZD} although some allowances made here to
ease this restrictiveness. For example, the literal \code{T} is optional,
fractional seconds are optional, and the time-zone designation (TZD) is only
required if the \code{long} and \code{full} standardized forms are requested
in \code{width}}

\item{locale}{the output locale for the printing of the date-time string.}

\item{width}{The standardized output date-time format, which can be one of
either \code{short}, \code{medium}, \code{long}, or \code{full}.}
}
\description{
Get a localized date-time in a standardized form.
}
