% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtpois.R
\name{dtpois}
\alias{dtpois}
\title{Right-Truncated Poisson Probability Mass Function}
\usage{
dtpois(y, lambda, c)
}
\arguments{
\item{y}{The vector of counts.}

\item{lambda}{The vector of means}

\item{c}{The cutoff of inflation}
}
\description{
PMF of a right truncated Poisson distribution.
}
\author{
Michael Floren
}
