% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_months.R
\name{first_day_of_week}
\alias{first_day_of_week}
\title{Get a named vector of all first-day-of-the-week names for different regions}
\usage{
first_day_of_week()
}
\value{
A character vector of short month names.
}
\description{
The \code{names_months()} function produces a vector of all short month names
used by the \strong{bigD} package.
}
\examples{
# Let's get a vector of regions where the
# first day of the week is Saturday
names(first_day_of_week()[first_day_of_week() == "sat"])

}
