% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdf_ftf.R
\name{flex_t12_vec}
\alias{flex_t12_vec}
\title{Get a vector of all 12-hour flexible time types}
\usage{
flex_t12_vec()
}
\value{
A character vector of supported 12-hour flexible time types.
}
\description{
The \code{flex_t12_vec()} function produces a vector of all supported flexible
12-hour time types in the \strong{bigD} package. These types are essentially
identifiers for classes of cross-locale time formatting, so, none of these
should be used directly in the \code{format} argument of the \code{fdt()} function (use
the \link{flex_t12_lst} object for that).
}
