% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols_patterns.R
\name{get_currency_symbol}
\alias{get_currency_symbol}
\title{Get the locale-specific form of a currency symbol}
\usage{
get_currency_symbol(name, locale, type = NULL)
}
\arguments{
\item{name}{The ISO 4217:2015 alphabetic code (three-letter) that defines a
currency. Examples include \code{"USD"} for U.S. Dollars and \code{"JPY"} for
Japanese Yen.}

\item{locale}{an optional output locale to use for formatting the \code{input}
value according to the specified locale's rules. Example locale names
include \code{"en_US"} for English (United States) and \code{"fr_FR"} for French
(France). If a locale isn't provided and certain require locale- based
formatting then the \code{"en_US"} locale is used for this purpose.}

\item{type}{The output representation of the currency symbol in the provided
\code{locale}. This can be \code{"symbol"}, \code{"narrow"}, or \code{"variant"}.}
}
\description{
Get the locale-specific form of a currency symbol
}
