% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_formatting.R
\name{insert_number_seps}
\alias{insert_number_seps}
\title{Provide separators to a numeric value with a grouping pattern}
\usage{
insert_number_seps(integer, grouping = c("333", "322"), sep_mark = ",")
}
\arguments{
\item{integer}{The integer portion of a numeric value. Can be supplied as an
integer value or a string. If supplied as a string, it must only contain
numeral characters.}

\item{grouping}{The grouping pattern of either "#,##0.###" with \code{"333"} or
"#,##,##0.###" with \code{"322"}.}

\item{sep_mark}{The character to use as the separator mark. By default, this
is the comma character.}
}
\description{
Provide separators to a numeric value with a grouping pattern
}
