% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_months.R
\name{names_months}
\alias{names_months}
\title{Get a vector of all the short month names}
\usage{
names_months()
}
\value{
A character vector of short month names.
}
\description{
The \code{names_months()} function produces a vector of all short month names
used by the \strong{bigD} package.
}
\examples{
# Let's get all the short month names with
# the `names_months()` function
names_months()

}
