% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_months.R
\name{names_wkdays}
\alias{names_wkdays}
\title{Get a vector of all the short weekday names}
\usage{
names_wkdays()
}
\value{
A character vector of short weekday names.
}
\description{
The \code{names_wkdays()} function produces a vector of all short weekday names
used by the \strong{bigD} package.
}
\examples{
# Let's get all the short weekday names with
# the `names_wkdays()` function
names_wkdays()

}
