% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_date_time.R
\name{standard_date_time}
\alias{standard_date_time}
\title{Obtain a standard datetime format that works across locales}
\usage{
standard_date_time(type = c("short", "medium", "long", "full"))
}
\arguments{
\item{type}{One of four standardized types for the resulting datetime that
range in precision and verbosity. These are \code{"short"} (the default),
\code{"medium"}, \code{"long"}, and \code{"full"}.}
}
\description{
Obtain a standard datetime format that works across locales
}
