% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_date_time.R
\name{standard_time}
\alias{standard_time}
\title{Obtain a standard time format that works across locales}
\usage{
standard_time(type = c("full", "long", "medium", "short"))
}
\description{
Obtain a standard time format that works across locales
}
