% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_dat.R
\name{tp_dat}
\alias{tp_dat}
\title{Generation of truncated Poisson data}
\usage{
tp_dat(n, lambda, c)
}
\arguments{
\item{n}{The number of data points to be generated.}

\item{lambda}{The mean parameter for the truncated Poisson.}

\item{c}{The cutoff of inflation.}
}
\description{
This function generates truncated Poisson data, with a truncation at c (counts can include c).
}
\details{
This generates data from a right truncated Poisson distribution, where generated y's are in the range \eqn{0 \le y \le c}.
}
\examples{
#show a histogram of right truncated Poisson data
hist(tp_dat(n=3000, lambda=3, c=5))

}
\author{
Michael Floren
}
