% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_formatting.R
\name{insert_number_seps}
\alias{insert_number_seps}
\title{For a non-decimal number string, insert separators
according to the grouping pattern}
\usage{
insert_number_seps(integer, grouping = c("#,##0.###", "#,##,##0.###"),
  sep_mark)
}
\arguments{
\item{integer}{the integer portion of a numeric vaue. Can be supplied as an
integer value or a string. If supplied as a string, it must only contain
numeral characters.}

\item{grouping}{the CLDR grouping pattern of either \code{"#,##0.###"} or
\code{"#,##,##0.###"}.}

\item{sep_mark}{the character to use as the separator mark.}
}
\description{
For a non-decimal number string, insert separators
according to the grouping pattern
}
