% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_date_time.R
\name{standard_date_time}
\alias{standard_date_time}
\title{Obtain a standard datetime format that works across locales}
\usage{
standard_date_time(type = c("full", "long", "medium", "short"))
}
\description{
Obtain a standard datetime format that works across locales
}
